ai-service/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
colorama==0.4.6
filelock==3.24.3
Flask==3.1.3
fsspec==2026.2.0
ftfy==6.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
regex==2026.2.28
requests==2.32.5
setuptools==82.0.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.6
